% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.r
\docType{data}
\name{all_hgnc_wtEntrez}
\alias{all_hgnc_wtEntrez}
\title{Map from Entrez to HGNC}
\format{
An object of class \code{"data frame"}.
}
\usage{
data(all_hgnc_wtEntrez)
}
\description{
Gene map from Entrez to HGNC, written by Nathan Skene et al. (2018) for their MAGMA.Celltyping package. Used to find rsIDs for SNPs when they're not present.
}
\examples{
data(all_hgnc_wtEntrez)
head(all_hgnc_wtEntrez)

}
\references{
1. Nathan Skene et al. Nat. Genet. 2018.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/29785013}{PubMed})
}
\keyword{entrez}
\keyword{gene}
\keyword{hgnc}
\keyword{map}
