% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_testing.r
\name{bin_specificity_into_quantiles}
\alias{bin_specificity_into_quantiles}
\title{Bin specificity into quantiles}
\usage{
bin_specificity_into_quantiles(input_matrix, numberOfBins)
}
\arguments{
\item{input_matrix}{Input matrix.}

\item{numberOfBins}{Number of bins.}
}
\value{
Dataframe
}
\description{
Mostly an adjuvant function for MAGMA linear, which requires quantiles (natural numbers) as input.
}
