% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_testing.r
\name{compute_bjk_enrichment_discrete}
\alias{compute_bjk_enrichment_discrete}
\title{Complete block jackknife joint test}
\usage{
compute_bjk_enrichment_discrete(
  gene_sets,
  sumstats_files,
  output_dir,
  pop = "eur",
  n_blocks = 200,
  number_of_threads = 1
)
}
\arguments{
\item{gene_sets}{List of genesets. Each geneset must be a character vector, containing HGNC gene names.}

\item{sumstats_files}{Addresses of properly formatted MAGMA and LDSC-formatted files for GWAS summary statistics.}

\item{output_dir}{Folder where you want the output to be stored. Default = NULL, which makes this function create a directory where results will be temporarily stored. The directory will be deleted afterwards, if this parameter is null. The only output will be the return. Otherwise, both the files and the return are kept.}

\item{pop}{Word describing ancestry of the GWAS data. Relevant so that the proper LD panel will be used (default = eur; alternatives = afr, amr, eas, sas, subpop)}

\item{n_blocks}{The number of blocks to split the genome into (default = 200).}

\item{number_of_threads}{Number of threads to parallelize over.}
}
\value{
List
}
\description{
Compute GWAS enrichment in discrete lists of genes with the block jackknife joint test.
}
