% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_scoring_big_matrices.r
\name{compute_ewce_selection_environ}
\alias{compute_ewce_selection_environ}
\title{Compute top N\% EWCE gene-sets - from environment-stored expression data}
\usage{
compute_ewce_selection_environ(
  obj,
  gene_selection_cutoff = 10,
  number_of_threads = 1
)
}
\arguments{
\item{obj}{Environment object containing an exp (numeric matrix of genes-by-cell-ID) & an annot (dataframe of cell annotations).}

\item{gene_selection_cutoff}{Selection cutoff as a percentile. Default = 10.}

\item{number_of_threads}{Number of threads to parallelize calculation over. Default = 1.}
}
\value{
List
}
\description{
Compute top N\% EWCE ranked genes for each cell-type. Output is a list of named character vectors, where each character vector represents a cell-type.
}
