% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_scoring_big_matrices.r
\name{compute_t_stat_scores_environ}
\alias{compute_t_stat_scores_environ}
\title{Compute t-statistic scores - from environment-stored expression data}
\usage{
compute_t_stat_scores_environ(
  obj,
  number_of_threads = 1,
  col_max_miss = 0.1,
  row_max_miss = 0.1
)
}
\arguments{
\item{obj}{Environment object containing an exp (numeric matrix of genes-by-cell-ID) & an annot (dataframe of cell annotations).}

\item{number_of_threads}{Number of threads to parallelize calculation over. Default = 1.}

\item{col_max_miss}{Column maximum missingness (default = 0.1)}

\item{row_max_miss}{Row maximum missingness (default = 0.1)}
}
\value{
Matrix
}
\description{
Compute gene-by-celltype specificity matrix with the t-statistic, then replaces the NA values for each row and finally scales the matrix between 0 and 1.
}
