% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dependencies.r
\name{dep_path}
\alias{dep_path}
\title{Dependency data paths}
\usage{
dep_path(
  pop = c("eur", "afr", "amr", "eas", "sas")[1],
  for_which = c("all", "magma", "snpsea", "ldsc")[1]
)
}
\arguments{
\item{pop}{Reference population shortcode (available right now: eur; Soon to be added: afr, amr, eas, eur, sas).}

\item{for_which}{Which softwares to check (available right now: all, magma, snpsea or ldsc).}
}
\value{
Named list (of paths)
}
\description{
Checks package-specific paths for where the dependency-data is usually stored. Dependency data is mostly data on LD regions, but other required stats too, and are usually population- (e.g. eas, afr, eur, etc.) and package- (LDSC, MAGMA & SNPsea) specific.
}
