% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dependencies.r
\name{find_magma_version}
\alias{find_magma_version}
\title{Find version of MAGMA binary}
\usage{
find_magma_version(path)
}
\arguments{
\item{path}{Path to MAGMA binary}
}
\value{
version (character)
}
\description{
Returns the version number of a MAGMA binary
}
