% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_scoring_big_matrices.r
\name{format_celltype_names}
\alias{format_celltype_names}
\title{Format celltype names}
\usage{
format_celltype_names(character_to_process)
}
\arguments{
\item{character_to_process}{Vector of characters.}
}
\value{
Character vector
}
\description{
Properly formats common symbols for MAGMA, LDSC & SNPsea.
}
