% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_testing.r
\name{gene_object_renamer}
\alias{gene_object_renamer}
\title{Gene object renamer}
\usage{
gene_object_renamer(gobj)
}
\arguments{
\item{gobj}{Either: 1. named list of character vectors (each vector is a gene-set, containing gene IDs), or 2. numeric matrix (each row is a gene, each column is a condition/group/set)}
}
\value{
List
}
\description{
Takes a named-list (of gene sets) or a matrix (of gene ranking sets) and renames the sets (e.g. from "CD4+ T-cell" to "CD4_T_cell")
}
