% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_scoring_big_matrices.r
\name{generate_celltype_data_environ}
\alias{generate_celltype_data_environ}
\title{Compute EWCE matrix - from environment-stored expression data}
\usage{
generate_celltype_data_environ(obj, number_of_threads = 1)
}
\arguments{
\item{obj}{Environment object containing an exp (numeric matrix of genes-by-cell-ID) & an annot (dataframe of cell annotations).}

\item{number_of_threads}{Number of threads to parallelize calculation over. Default = 1.}
}
\value{
Dataframe
}
\description{
Compute gene-by-celltype specificity matrix with the EWCE score. Code modified from the EWCE package by Skene et al. (2016, DOI 10.3389/fnins.2016.00016).
}
