% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_testing.r
\name{get_genomebuild_for_sumstats}
\alias{get_genomebuild_for_sumstats}
\title{Get genome build from sumstats}
\usage{
get_genomebuild_for_sumstats(path)
}
\arguments{
\item{path}{Path of summary statistics file.}
}
\value{
Character
}
\description{
Tries to estimate what genome build a sumstats file comes from.
}
