% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_testing.r
\name{get_magma_paths}
\alias{get_magma_paths}
\title{Get MAGMA paths}
\usage{
get_magma_paths(
  sumstats_file,
  upstream_kb,
  downstream_kb,
  population,
  gene_nomenclature
)
}
\arguments{
\item{sumstats_file}{Path for summary statistics file.}

\item{upstream_kb}{Kilobases upstream of gene to add.}

\item{downstream_kb}{Kilobases downstream of gene to add.}

\item{population}{Population (default = eur)}

\item{gene_nomenclature}{Gene nomenclature (default = hgnc)}
}
\value{
List
}
\description{
Get or create MAGMA folders for the sumstats file provided. Outputs a path.
}
