% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dependencies.r
\name{get_os}
\alias{get_os}
\title{Detect OS}
\usage{
get_os()
}
\value{
character vector of size 1
}
\description{
Function that outputs "windows", "linux" or "osx" depending on what system it is run on. Created by Will from R-bloggers here: https://www.r-bloggers.com/2015/06/identifying-the-os-from-r/
}
