% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwascelltyper.r
\docType{package}
\name{gwascelltyper}
\alias{gwascelltyper}
\title{gwascelltyper}
\description{
Fast GWAS enrichment testing, using LDSC, SNPsea & MAGMA (and code from MAGMA.Celltyping by N. Skene et al. 2018).
}
\author{
Alexandru-Ioan Voda \email{alexandru.voda@seh.ox.ac.uk}
}
