% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_scoring_big_matrices.r
\name{make_geneset_from_score_matrix}
\alias{make_geneset_from_score_matrix}
\title{Make gene-set from gene-by-celltype specificity matrix}
\usage{
make_geneset_from_score_matrix(gene_score_matrix, gene_selection_cutoff = 10)
}
\arguments{
\item{gene_score_matrix}{Gene-by-cell-type specificity matrix.}

\item{gene_selection_cutoff}{Selection cutoff as a percentile. Default = 10.}
}
\value{
List
}
\description{
Returns list of the top N\% (largest values) genes for each cell-type. Output is a list of named character vectors, where each character vector represents a cell-type.
}
