% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_testing.r
\name{map_SNPs_to_genes_for_MAGMA}
\alias{map_SNPs_to_genes_for_MAGMA}
\title{Map SNPs to genes for MAGMA}
\usage{
map_SNPs_to_genes_for_MAGMA(
  sumstats_file,
  upstream_kb = 10,
  downstream_kb = 1.5,
  N = NULL,
  pop = "eur",
  gene_nomenclature = "hgnc",
  deps = dep_path(pop, "magma")
)
}
\arguments{
\item{sumstats_file}{Path to summary statistics file.}

\item{upstream_kb}{Default = 10}

\item{downstream_kb}{Default = 1.5}

\item{N}{Default = NULL}

\item{pop}{Default = eur}

\item{gene_nomenclature}{Default = hgnc}

\item{deps}{A named list, containing all of the dependencies for MAGMA (e.g. executables, SNP linkage disequilibrium data, etc.). Provided by default by another function, called gwascelltyper::dep_path().}
}
\value{
Character
}
\description{
Compute GWAS enrichment in discrete lists of genes with MAGMA.
}
