% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sumstats.r
\name{munge_sumstats_for_LDSC_from_MAGMA}
\alias{munge_sumstats_for_LDSC_from_MAGMA}
\title{Process summary statistics for LDSC}
\usage{
munge_sumstats_for_LDSC_from_MAGMA(
  sumstats_magma = NULL,
  sumstats_munged = paste0(sumstats_magma, "_processed.sumstats.gz"),
  number_of_threads = 1,
  gwas_sample_size = NULL,
  merge_alleles = paste0(system.file(package = "gwascelltyper"),
    "/extdata/w_hm3.snplist"),
  ldsc_path = paste0(system.file(package = "gwascelltyper"), "/extdata/"),
  preview = FALSE
)
}
\arguments{
\item{sumstats_magma}{File containing the raw summary statistics file.}

\item{sumstats_munged}{File to output into. Default = paste0(sumstats_raw, "_processed.sumstats").}

\item{number_of_threads}{Number of threads to parallelize calculation over. Default = 1.}

\item{gwas_sample_size}{If GWAS sample size is not a constant integer for all SNPs, then leave NULL. Otherwise, must be an integer.}

\item{merge_alleles}{Leave NULL if you do not want to merge the SNPs with those in HapMap3 (as recommended in LDSC guidelines). Otherwise, should be the path of the w_hm3.snplist file.}

\item{ldsc_path}{Path to LDSC python script (default is extdata folder in package path).}

\item{preview}{Show ideal LDSC sumstats file then exit. If false (default), it proceeds to the processing.}
}
\value{
Path
}
\description{
Process genome-wide association study summary statistics file into the required LDSC format (given MAGMA-formatted sumstats) for analyses.
}
