% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sumstats.r
\name{munge_sumstats_for_MAGMA}
\alias{munge_sumstats_for_MAGMA}
\title{Process summary statistics for MAGMA}
\usage{
munge_sumstats_for_MAGMA(
  sumstats_raw,
  sumstats_munged = paste0(sumstats_raw, "_processed.sumstats"),
  preview = FALSE
)
}
\arguments{
\item{sumstats_raw}{File containing the raw summary statistics file.}

\item{sumstats_munged}{File to output into. Default = paste0(sumstats_raw, "_processed.sumstats").}

\item{preview}{Show ideal MAGMA sumstats file then exit. If false (default), it proceeds to the processing.}
}
\value{
Path
}
\description{
Process genome-wide association study summary statistics file into the required MAGMA format for analyses.
}
