% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sumstats.r
\name{munge_sumstats_for_SNPSEA_from_MAGMA}
\alias{munge_sumstats_for_SNPSEA_from_MAGMA}
\title{Process summary statistics for SNPsea}
\usage{
munge_sumstats_for_SNPSEA_from_MAGMA(
  sumstats_magma = NULL,
  sumstats_munged = paste0(sumstats_magma, "_processed.sumstats"),
  preview = FALSE
)
}
\arguments{
\item{sumstats_magma}{File containing GWAS summary statistics file properly formatted for MAGMA.}

\item{sumstats_munged}{File to output SNPsea-format into. Default = paste0(sumstats_magma, "_processed.sumstats").}

\item{preview}{Show ideal SNPsea sumstats file then exit. If false (default), it proceeds to the processing.}
}
\value{
Path
}
\description{
Process genome-wide association study summary statistics file into the required SNPsea format (given MAGMA-formatted sumstats) for analyses. Leaves only the genome-wide significant SNPs (because that is the SNPsea approach).
}
