% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_scoring_big_matrices.r
\name{range01}
\alias{range01}
\title{Scale matrix between 0 and 1.}
\usage{
range01(m)
}
\arguments{
\item{m}{Numeric matrix}
}
\value{
Matrix
}
\description{
Scales matrix by min and max between 0 and 1. Code from https://stackoverflow.com/a/5665644. Adjuvant function.
}
