% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_scoring_big_matrices.r
\name{sanity_check_for_exp_and_annot_environ}
\alias{sanity_check_for_exp_and_annot_environ}
\title{Sanity checking exp-annot correspondence}
\usage{
sanity_check_for_exp_and_annot_environ(
  obj,
  format_cell_names = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{obj}{Environment to read exp and annotLevels to.}

\item{format_cell_names}{Prepare cell/sample names for MAGMA, LDSC, SNPsea (default = TRUE).}

\item{verbose}{Whether to print intermediate work (default = TRUE).}
}
\value{
None
}
\description{
Checking that exp and annot have 1-to-1 correspondence. Eliminating excesive/needless annots.
}
