% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.r
\docType{data}
\name{snp_loc}
\alias{snp_loc}
\title{SNP location data}
\format{
An object of class \code{"data frame"}.
}
\usage{
data(snp_loc)
}
\description{
SNP location written by Nathan Skene et al. (2018) for their MAGMA.Celltyping package. Used to find rsIDs for SNPs when they're not present.
}
\examples{
data(standard_sumstats_column_headers)
head(standard_sumstats_column_headers)

}
\references{
1. Nathan Skene et al. Nat. Genet. 2018.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/29785013}{PubMed})
}
\keyword{gwas}
\keyword{location}
\keyword{snp}
