% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.r
\docType{data}
\name{standard_sumstats_column_headers}
\alias{standard_sumstats_column_headers}
\title{Standard summstats column headers}
\format{
An object of class \code{"data frame"}; see \code{\link[data.frame]{data.frame}}.
}
\usage{
data(standard_sumstats_column_headers)
}
\description{
Equivalency table written by Nathan Skene et al. (2018) for their MAGMA.Celltyping package. Used to standardize GWAS summary statistics headers.
}
\examples{
data(standard_sumstats_column_headers)
head(standard_sumstats_column_headers)

}
\references{
1. Nathan Skene et al. Nat. Genet. 2018.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/29785013}{PubMed})
}
\keyword{column}
\keyword{gwas}
\keyword{headers}
\keyword{standard}
\keyword{sumstats}
