% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_testing.r
\name{compute_LDSC_enrichment_discrete_optimized}
\alias{compute_LDSC_enrichment_discrete_optimized}
\title{Compute LDSC enrichment - optimized}
\usage{
compute_LDSC_enrichment_discrete_optimized(
  gene_sets,
  sumstats_file,
  output_dir = NULL,
  number_of_threads = 1,
  window_size = 1e+05,
  pop = "eur",
  ld_wind_cm = 1,
  deps = dep_path(pop, "ldsc")
)
}
\arguments{
\item{gene_sets}{List of genesets. Each geneset must be a character vector, containing HGNC gene names.}

\item{sumstats_file}{Address of properly formatted LDSC GWAS file.}

\item{output_dir}{Folder where you want the output to be stored. Default = NULL, which makes this function create a directory where results will be temporarily stored. The directory will be deleted afterwards, if this parameter is null. The only output will be the return. Otherwise, both the files and the return are kept.}

\item{number_of_threads}{Number of threads to parallelize calculation over. Default = 1.}

\item{window_size}{Size of window surrounding gene, for SNP-to-gene mapping purposes. Default = 100000 (100kb).}

\item{pop}{Word describing ancestry of the GWAS data. Relevant so that the proper LD panel will be used (default = eur; alternatives = afr, amr, eas, sas, subpop)}

\item{ld_wind_cm}{Specify the window size to be used for estimating LD Scores in units of centiMorgans (cM). Default = 1.}

\item{deps}{A named list, containing all of the dependencies for LDSC (e.g. Python script paths, SNP linkage disequilibrium data, etc.). Provided by default by another function, called gwascelltyper::dep_path().}

\item{gene_nomenclature}{Type of gene IDs fed in from the expression data (default = hgnc; alternatives = entrez).}
}
\value{
Dataframe
}
\description{
Compute GWAS enrichment in discrete lists of genes with LDSC. Optimized to load all dependency files into Python-allocated RAM only once.
}
