% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_testing.r
\name{compute_MAGMA_enrichment_linear}
\alias{compute_MAGMA_enrichment_linear}
\title{Compute MAGMA enrichment linear}
\usage{
compute_MAGMA_enrichment_linear(
  gene_score_matrix,
  sumstats_file,
  output_dir = NULL,
  upstream_kb = 10,
  downstream_kb = 1.5,
  gene_nomenclature = "hgnc",
  pop = "eur",
  gwas_sample_size,
  deps = dep_path(pop, "magma")
)
}
\arguments{
\item{gene_score_matrix}{Matrix with HGNC genes as rows & cell-types as columns. Values need to be numeric.}

\item{sumstats_file}{Address of properly formatted MAGMA GWAS file.}

\item{output_dir}{Folder where you want the output to be stored. Default = NULL, which makes this function create a directory where results will be temporarily stored. The directory will be deleted afterwards, if this parameter is null. The only output will be the return. Otherwise, both the files and the return are kept.}

\item{upstream_kb}{Number of kb upstream of gene, SNP-to-gene mapping window. Default = 10.}

\item{downstream_kb}{Number of kb downstream of gene, SNP-to-gene mapping window. Default = 1.5.}

\item{gene_nomenclature}{(default = hgnc)}

\item{pop}{Word describing ancestry of the GWAS data. Relevant so that the proper LD panel will be used (default = eur; alternatives = afr, amr, eas, sas, subpop)}

\item{gwas_sample_size}{Sample size of GWAS.}

\item{deps}{A named list, containing all of the dependencies for MAGMA (e.g. executables, SNP linkage disequilibrium data, etc.). Provided by default by another function, called gwascelltyper::dep_path().}
}
\value{
Dataframe
}
\description{
Compute GWAS enrichment in gene specificity matrices (linear analysis) with MAGMA.
}
