% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dependencies.r
\name{download_dependencies}
\alias{download_dependencies}
\title{Download dependencies}
\usage{
download_dependencies(
  address = paste0(system.file(package = "gwascelltyper"), "/extdata/"),
  pop = c("eur", "eas", "afr", "amr", "sas", "subpop")[1]
)
}
\arguments{
\item{address}{Path where dependencies are downloaded. Do not modify, parameter meant for developers.}

\item{pop}{Population to download LD data for (available right now: eur; available soon: eas, afr, amr, sas, subpop).}
}
\value{
Path
}
\description{
Downloads data dependencies (e.g. linkage disequilibrium, gene annotations, etc.) pre-formatted for each of the 3 packages (LDSC, MAGMA & SNPsea)
}
