% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_testing.r
\name{map_SNPs_to_genes_for_MAGMA}
\alias{map_SNPs_to_genes_for_MAGMA}
\title{Map SNPs to genes for MAGMA}
\usage{
map_SNPs_to_genes_for_MAGMA(
  sumstats_file,
  upstream_kb = 10,
  downstream_kb = 1.5,
  N = NULL,
  population = "eur",
  gene_nomenclature = "hgnc",
  genome_ref_path = paste0(system.file(package = "gwascelltyper"), "/extdata/g1000_",
    population),
  magma_path = paste0(system.file(package = "gwascelltyper"), "/extdata/magma-linux64")
)
}
\arguments{
\item{sumstats_file}{Path to summary statistics file.}

\item{upstream_kb}{Default = 10}

\item{downstream_kb}{Default = 1.5}

\item{N}{Default = NULL}

\item{population}{Default = eur}

\item{gene_nomenclature}{Default = hgnc}

\item{genome_ref_path}{Path to g1000 dependency files. Depends on population used.}

\item{magma_path}{Path to MAGMA executable.}
}
\value{
Character
}
\description{
Compute GWAS enrichment in discrete lists of genes with MAGMA.
}
